Push map strategy logic into map card
