+++ src/panels/lovelace/cards/hui-map-card.ts
@@ -12,6 +12,7 @@ import memoizeOne from "memoize-one";
 import { getColorByIndex } from "../../../common/color/colors";
 import { isComponentLoaded } from "../../../common/config/is_component_loaded";
 import { computeDomain } from "../../../common/entity/compute_domain";
+import { computeStateDomain } from "../../../common/entity/compute_state_domain";
 import { computeStateName } from "../../../common/entity/compute_state_name";
 import { deepEqual } from "../../../common/util/deep-equal";
 import parseAspectRatio from "../../../common/util/parse-aspect-ratio";
@@ -80,14 +81,38 @@ class HuiMapCard extends LitElement implements LovelaceCard {
 
   private _subscribed?: Promise<(() => Promise<void>) | undefined>;
 
+  private _getAllEntities(): string[] {
+    const hass = this.hass!;
+    const personSources = new Set<string>();
+    const locationEntities: string[] = [];
+    Object.values(hass.states).forEach((entity) => {
+      if (
+        !("latitude" in entity.attributes) ||
+        !("longitude" in entity.attributes)
+      ) {
+        return;
+      }
+      locationEntities.push(entity.entity_id);
+      if (computeStateDomain(entity) === "person" && entity.attributes.source) {
+        personSources.add(entity.attributes.source);
+      }
+    });
+
+    return locationEntities.filter((entity) => !personSources.has(entity));
+  }
+
   public setConfig(config: MapCardConfig): void {
     if (!config) {
       throw new Error("Error in card configuration.");
     }
 
-    if (!config.entities?.length && !config.geo_location_sources) {
+    if (
+      !config.show_all &&
+      !config.entities?.length &&
+      !config.geo_location_sources
+    ) {
       throw new Error(
-        "Either entities or geo_location_sources must be specified"
+        "Either show_all, entities, or geo_location_sources must be specified"
       );
     }
     if (config.entities && !Array.isArray(config.entities)) {
@@ -101,9 +126,12 @@ class HuiMapCard extends LitElement implements LovelaceCard {
     }
 
     this._config = config;
-    this._configEntities = config.entities
-      ? processConfigEntities<MapEntityConfig>(config.entities)
-      : [];
+    this._configEntities =
+      this.hass && config.show_all
+        ? processConfigEntities<MapEntityConfig>(this._getAllEntities())
+        : config.entities
+          ? processConfigEntities<MapEntityConfig>(config.entities)
+          : [];
     this._mapEntities = this._getMapEntities();
   }
 
@@ -310,6 +338,18 @@ class HuiMapCard extends LitElement implements LovelaceCard {
     }
     if (changedProps.has("_config")) {
       this._computePadding();
+    }
+
+    if (
+      this._config?.show_all &&
+      this._configEntities?.length === 0 &&
+      changedProps.has("hass") &&
+      changedProps.get("hass") === undefined
+    ) {
+      this._configEntities = processConfigEntities<MapEntityConfig>(
+        this._getAllEntities()
+      );
+      this._mapEntities = this._getMapEntities();
     }
   }
